+++ src/themes/bootstrap5.js
@@ -40,6 +40,8 @@ class ThemeBootstrap5 extends Theme {
     const jsonData = super.getJsonData(config)
     jsonData.control.classList.add('mb-3')
     jsonData.input.classList.add('form-control')
+    jsonData.copyBtn.classList.add('btn-secondary')
+    jsonData.copyBtn.classList.add('w-100')
     jsonData.saveBtn.classList.add('btn-primary')
     jsonData.saveBtn.classList.add('w-100')
     return jsonData